Update and rename src/plugins/example.js to src/plugins/example/index.js
