--- src/plugins/example.js
@@ -1,79 +1,0 @@
-'use strict';
-var plugin = [];
-plugin.name = ""; // Name of plugin where it would be listed in
-plugin.author = "andrews54757"; // author
-plugin.command = []; // extra commands
-plugin.commandName = []; // extra commands
-plugin.addToHelp = [];
-plugin.commandName[0] = "tptoplayer";
-plugin.addToHelp[0] = "tptoplayer : tps to a player";
-plugin.gamemodeId = []; // todo with plugin gamemodes
-plugin.gamemode = []; // todo with plugin gamemodea
-plugin.version = '1.0.0';
-plugin.compatVersion = ''; // compatable with (todo)
-plugin.init = function (gameServer) {
-  // init, Used to do stuff such as overriding things
-  
-  
-  
-}
-plugin.command[0] = function(gameServer,split) {
-
- var id = parseInt(split[1]);
- var idt = parseInt(split[2]);
-  if (isNaN(id)) {
-    console.log("[Console] Please specify a valid player ID!");
-    return;
-  }
-  if (isNaN(idt)) {
-    console.log("[Console] Please specify a valid target player ID!");
-    return;
-  }
-  var pos;
-  var ok = false;
-  var target;
-  for (var i in gameServer.clients) {
-    if (gameServer.clients[i].playerTracker.pID == idt) {
-      var client = gameServer.clients[i].playerTracker;
-
-      for (var j in client.cells) {
-        client.cells[j].mass = 1000;
-        var x = fillChar(client.centerPos.x >> 0, ' ', 5, true);
-        var y = fillChar(client.centerPos.y >> 0, ' ', 5, true);
-
-        pos = {
-          x: x,
-          y: y
-        };
-      }
-      ok = true;
-      target = client
-      break;
-    }
-  }
-  if (!ok) {
-    console.log("[Console] Invalid Player!");
-    return;
-  }
-  // Make sure the input values are numbers
-  if (isNaN(pos.x) || isNaN(pos.y)) {
-    console.log("[Console] Invalid coordinates");
-    return;
-  }
-
-  // Spawn
-  for (var i in gameServer.clients) {
-    if (gameServer.clients[i].playerTracker.pID == id) {
-      var client = gameServer.clients[i].playerTracker;
-      for (var j in client.cells) {
-        client.cells[j].position.x = pos.x;
-        client.cells[j].position.y = pos.y;
-      }
-
-      console.log("[Console] Teleported " + client.name + " to " + target.name);
-      break;
-    }
-  }
-};
-module.exports = plugin;
-
